backend/server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
requests>=2.31.0
groq==0.4.1
pydantic==2.12.3
